requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
torch>=2.1
rdkit-pypi>=2022.9.5
transformers>=4.40
scikit-learn>=1.3
kagglehub[pandas-datasets]>=0.2.7
